Add addDriverSlots API call
API call for addDriverSlots will return any driverSlots records
created.  This also adds a UNIQUE index on driverSlots to driverId
and slotId.  If you try to add a driverSlot record where it already
exists it will simply not add it and log a duplicate INSERT error
but not prevent it from adding any other missing driverSlots.