Add taps/kegs functionality
When a customer places an order we look at the category of each line
item chosen and if it is a Small Keg or Large Keg then we increment the
kegs and taps quantity by one. There are possibly other ways to check if
a product for a line item is a Keg or note and I added notes about this
in the controller code.

When a driver picks up kegs from the customer to be returned to the
stores they record how many taps and kegs were returned. At this point
in time we are not tracking specific kegs, simply the number that were
returned. It will be up to the stores to reconcile their keg inventory.